requirements.txt
flask==2.3.3
rembg==2.0.38
pillow==10.0.0
gunicorn==21.2.0